VeryBigSad-moex-hack-main/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.6"
fastapi = "^0.104.1"
bcrypt = "^4.1.1"
pydantic-settings = "^2.1.0"
gunicorn = "^21.2.0"
uvicorn = "^0.24.0.post1"
tortoise-orm = "^0.20.0"
asyncpg = "^0.29.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic = {extras = ["email"], version = "^2.5.2"}
moexalgo = "^1.0.1"
pandas = "^2.1.3"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
